Add a list of known friend codes and blacklist them
This imports a list of friend codes currently blacklisted on Bruteforce
Movable.

The list of friend codes seen in blacklist.py are shown often in video
guides or viewed when a friendbot's friend code appears somewhere. In
neither of these cases should these be allowed to be mined, as it is
known for a fact that these are not the end user's friend code.